Feat: CurrenciesList: move orchestration to hook
