+++ src/components/Currency/CurrenciesList.tsx
@@ -14,28 +14,7 @@ import { currenciesActions, fetchCurrencies } from 'components/Currency/redux/cu
 import { searchSelector } from 'components/SearchBar/redux/searchBarSelectors';
 
 function CurrenciesList() {
-    const dispatch = useAppDispatch();
-    const { state: loadingStatus } = useAppSelector(getLoadingStatus);
-    const { searchField } = useAppSelector(searchSelector);
-    const filteredCurrencies = useAppSelector(getFilteredCurrencies);
-    const baseCurrency = useAppSelector(getBaseCurrency);
-    const currencyItems = filteredCurrencies.length ? (
-        filteredCurrencies.map((currency) => (
-            // React gives a false warning about this line that this shouldn't be wrapped in a <div>
-            <CurrencyItem key={currency.currency} baseCurrency={baseCurrency} currencyData={currency} />
-        ))
-    ) : (
-        <></>
-    );
-    useEffect(() => {
-        if (loadingStatus === apiState.PRISTINE) {
-            dispatch(fetchCurrencies());
-        }
-    }, [loadingStatus, dispatch]);
-
-    useEffect(() => {
-        dispatch(currenciesActions.updateFilteredCurrencies(searchField));
-    }, [searchField, dispatch]);
+    const { loadingStatus, filteredCurrencies, currencyItems } = useGetCurrenciesListData();
     return (
         <>
             {loadingStatus === apiState.LOADING && <div>Loading Currencies...</div>}
